Move all TensorPipe code in single file, avoiding preprocessor flag
Differential Revision: [D22840087](https://our.internmc.facebook.com/intern/diff/D22840087/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D22840087/)!

[ghstack-poisoned]
