+++ torch/csrc/distributed/rpc/tensorpipe_agent.cpp
@@ -159,6 +159,207 @@ C10_REGISTER_CREATOR(
 
 } // namespace
 
+///////////////////////  Serialization/Deserialization  ///////////////////////
+
+namespace {
+
+// The TensorPipe agent splits the RPC message's information across multiple
+// payloads. This allows the agent to provide the data to TensorPipe without
+// performing a copy into a single contiguous buffer, and without storing it as
+// metadata, which is less efficient.
+
+// First come the rpc::Message::type() and ::id().
+constexpr int kTpMessageTypeIdx = 0;
+constexpr int kTpMessageIdIdx = 1;
+// Then comes the rpc::Message::payload();
+constexpr int kTpMessagePayloadIdx = 2;
+// Last comes the pickle of rpc::Message::tensors() (with the tensors themselves
+// stored as, well, tensors in the tensorpipe::Message).
+constexpr int kTpMessagePickleIdx = 3;
+
+// We use vector<char> as the type of blobs because it's what rpc::Message uses
+// for its payload, even though it has the disadvantage that it cannot be
+// allocated with uninitialized memory: it is always zeroed out.
+
+// A struct that holds pointers that keep alive all the memory that will be
+// accessed by TensorPipe during a write operation.
+struct TensorpipeWriteBuffers {
+  // Allocate on heap so pointers stay valid as we move the holder.
+  std::unique_ptr<MessageType> type;
+  std::unique_ptr<int64_t> id;
+  std::vector<char> payload;
+  std::vector<char> pickle;
+  // This contains the original tensors and the clones of the sparse tensors.
+  std::vector<torch::Tensor> tensors;
+  // This contains the copies of the data of the tensors that didn't own their
+  // memory, e.g., the ones created from torch::from_blob() with no deleter.
+  std::vector<std::vector<char>> copiedTensors;
+};
+
+// A struct that holds pointers that keep alive all the memory that will be
+// accessed by TensorPipe during a read operation.
+struct TensorpipeReadBuffers {
+  // Allocate on heap so pointers stay valid as we move the holder.
+  std::unique_ptr<MessageType> type;
+  std::unique_ptr<int64_t> id;
+  std::vector<char> payload;
+  std::vector<char> pickle;
+  std::vector<c10::DataPtr> tensors;
+};
+
+// Convert an RPC message into a TensorPipe message, plus a holder to all the
+// data that must be kept alive while the write is performed asynchronously.
+std::tuple<tensorpipe::Message, TensorpipeWriteBuffers> tensorpipeSerialize(
+    Message&& rpcMessage) {
+  tensorpipe::Message tpMessage;
+  TensorpipeWriteBuffers buffers;
+
+  // Metadata
+  buffers.type = std::make_unique<MessageType>(rpcMessage.type());
+  buffers.id = std::make_unique<int64_t>(rpcMessage.id());
+  tpMessage.payloads.push_back(
+      tensorpipe::Message::Payload{buffers.type.get(), sizeof(MessageType)});
+  tpMessage.payloads.push_back(
+      tensorpipe::Message::Payload{buffers.id.get(), sizeof(int64_t)});
+
+  // Payload
+  buffers.payload = std::move(rpcMessage.payload());
+  // TensorPipe uses the same Message class for both reading and writing, thus
+  // it uses non-const pointers even though it doesn't modify them when writing.
+  // NOLINTNEXTLINE(cppcoreguidelines-pro-type-const-cast)
+  char* payloadPtr = const_cast<char*>(buffers.payload.data());
+  tpMessage.payloads.push_back(
+      tensorpipe::Message::Payload{payloadPtr, buffers.payload.size()});
+
+  // Tensors
+  buffers.tensors = cloneSparseTensors(rpcMessage.tensors()).vec();
+  torch::jit::Pickler pickler([&](const void* buf, size_t sz) -> size_t {
+    buffers.pickle.insert(
+        buffers.pickle.end(),
+        static_cast<const char*>(buf),
+        static_cast<const char*>(buf) + sz);
+    return sz;
+  });
+  pickler.protocol();
+  pickler.pushIValue(buffers.tensors);
+  pickler.stop();
+  tpMessage.payloads.push_back(tensorpipe::Message::Payload{
+      buffers.pickle.data(), buffers.pickle.size()});
+  for (const auto& tensor : pickler.tensorData()) {
+    const auto& tensorData = jit::getWriteableTensorData(tensor);
+    // Enforce memory copy if tensor is created from torch::from_blob, means
+    // that the tensor doesn't own the memory.
+    if (!tensorData.storageHasDeleter()) {
+      std::vector<char> storageData(
+          tensorData.data(), tensorData.data() + tensorData.sizeInBytes());
+      tpMessage.tensors.push_back(
+          tensorpipe::Message::Tensor{storageData.data(), storageData.size()});
+      buffers.copiedTensors.push_back(std::move(storageData));
+    } else {
+      // TensorPipe uses the same Message class for both reading and writing, so
+      // it uses non-const ptrs even though it doesn't modify them when writing.
+      // NOLINTNEXTLINE(cppcoreguidelines-pro-type-const-cast)
+      char* tensorPtr = const_cast<char*>(tensorData.data());
+      tpMessage.tensors.push_back(
+          tensorpipe::Message::Tensor{tensorPtr, tensorData.sizeInBytes()});
+    }
+  }
+
+  return std::make_tuple(std::move(tpMessage), std::move(buffers));
+}
+
+// Allocate the buffers that will hold the incoming data. They will be managed
+// by the returned holder, which must be kept alive until the asynchronous read
+// has finished. Pointers to these buffers will be stored in-place in the
+// TensorPipe message.
+TensorpipeReadBuffers tensorpipeAllocate(tensorpipe::Message& tpMessage) {
+  TensorpipeReadBuffers buffers;
+
+  TORCH_INTERNAL_ASSERT(
+      tpMessage.payloads.size() == 4,
+      "message expected to contain 4 payloads, whereas it contained ",
+      tpMessage.payloads.size(),
+      " payloads");
+
+  TORCH_INTERNAL_ASSERT(
+      tpMessage.payloads[kTpMessageTypeIdx].length == sizeof(MessageType),
+      "first payload expected to contain ",
+      sizeof(MessageType),
+      " bytes, whereas it contained ",
+      tpMessage.payloads[kTpMessageTypeIdx].length,
+      " bytes");
+  buffers.type = std::make_unique<MessageType>();
+  tpMessage.payloads[kTpMessageTypeIdx].data = buffers.type.get();
+
+  TORCH_INTERNAL_ASSERT(
+      tpMessage.payloads[kTpMessageIdIdx].length == sizeof(int64_t),
+      "second payload expected to contain ",
+      sizeof(int64_t),
+      " bytes, whereas it contained ",
+      tpMessage.payloads[kTpMessageIdIdx].length,
+      " bytes");
+  buffers.id = std::make_unique<int64_t>();
+  tpMessage.payloads[kTpMessageIdIdx].data = buffers.id.get();
+
+  // FIXME The two resizes below zero out the vectors, which is not needed.
+
+  buffers.payload.resize(tpMessage.payloads[kTpMessagePayloadIdx].length);
+  tpMessage.payloads[kTpMessagePayloadIdx].data = buffers.payload.data();
+
+  buffers.pickle.resize(tpMessage.payloads[kTpMessagePickleIdx].length);
+  tpMessage.payloads[kTpMessagePickleIdx].data = buffers.pickle.data();
+
+  for (auto& tensor : tpMessage.tensors) {
+    buffers.tensors.push_back(at::getCPUAllocator()->allocate(tensor.length));
+    tensor.data = buffers.tensors.back().get();
+  }
+
+  return buffers;
+}
+
+// Convert a TensorPipe message back into an RPC message. This requires the data
+// to be available and can thus only be performed once the asynchronous read has
+// completed. The holder can be destroyed once this function returns.
+Message tensorpipeDeserialize(
+    tensorpipe::Message&& message,
+    TensorpipeReadBuffers&& buffers) {
+  // Tensors
+  std::vector<at::Tensor> tensors;
+  const char* pickleData = buffers.pickle.data();
+  size_t pickleLen = buffers.pickle.size();
+  size_t picklePos = 0;
+  auto pickleReadFunc = [&](char* buf, size_t n) -> size_t {
+    if (picklePos >= pickleLen || n == 0) {
+      return 0;
+    }
+    size_t toCopy = std::min(picklePos + n, pickleLen) - picklePos;
+    memcpy(buf, pickleData + picklePos, toCopy);
+    picklePos += toCopy;
+    return toCopy;
+  };
+  auto tensorReadFunc = [&](const std::string& ename) -> at::DataPtr {
+    unsigned long index = std::stoul(ename);
+    return std::move(buffers.tensors.at(index));
+  };
+
+  // No need to pass typeResolver here, as it always processes string and
+  // tensors only
+  torch::jit::Unpickler unpickler(
+      pickleReadFunc, nullptr, nullptr, tensorReadFunc, {});
+  auto ival = unpickler.parse_ivalue();
+  for (auto&& t : ival.toTensorList()) {
+    tensors.emplace_back(std::move(t));
+  }
+
+  return Message(
+      std::move(buffers.payload),
+      std::move(tensors),
+      *buffers.type,
+      *buffers.id);
+}
+
+}
+
 //////////////////////////  MetricsTracker  /////////////////////////////////
 
 TensorPipeAgent::TimeSeriesMetricsTracker::TimeSeriesMetricsTracker(
