+++ torch/csrc/distributed/rpc/utils.h
@@ -4,10 +4,6 @@
 #include <torch/csrc/jit/serialization/pickle.h>
 #include <torch/csrc/utils/byte_order.h>
 
-namespace tensorpipe {
-class Message;
-} // namespace tensorpipe
-
 namespace torch {
 namespace distributed {
 namespace rpc {
@@ -51,55 +47,6 @@ TORCH_API std::string wireSerialize(
 TORCH_API std::pair<std::vector<char>, std::vector<at::Tensor>> wireDeserialize(
     const void* data,
     size_t data_size);
-
-// We use vector<char> as the type of blobs because it's what rpc::Message uses
-// for its payload, even though it has the disadvantage that it cannot be
-// allocated with uninitialized memory: it is always zeroed out.
-
-// A struct that holds pointers that keep alive all the memory that will be
-// accessed by TensorPipe during a write operation.
-struct TensorpipeWriteBuffers {
-  // Allocate on heap so pointers stay valid as we move the holder.
-  std::unique_ptr<MessageType> type;
-  std::unique_ptr<int64_t> id;
-  std::vector<char> payload;
-  std::vector<char> pickle;
-  // This contains the original tensors and the clones of the sparse tensors.
-  std::vector<torch::Tensor> tensors;
-  // This contains the copies of the data of the tensors that didn't own their
-  // memory, e.g., the ones created from torch::from_blob() with no deleter.
-  std::vector<std::vector<char>> copiedTensors;
-};
-
-// A struct that holds pointers that keep alive all the memory that will be
-// accessed by TensorPipe during a read operation.
-struct TensorpipeReadBuffers {
-  // Allocate on heap so pointers stay valid as we move the holder.
-  std::unique_ptr<MessageType> type;
-  std::unique_ptr<int64_t> id;
-  std::vector<char> payload;
-  std::vector<char> pickle;
-  std::vector<c10::DataPtr> tensors;
-};
-
-// Convert an RPC message into a TensorPipe message, plus a holder to all the
-// data that must be kept alive while the write is performed asynchronously.
-TORCH_API std::tuple<tensorpipe::Message, TensorpipeWriteBuffers>
-tensorpipeSerialize(Message&& rpcMessage);
-
-// Allocate the buffers that will hold the incoming data. They will be managed
-// by the returned holder, which must be kept alive until the asynchronous read
-// has finished. Pointers to these buffers will be stored in-place in the
-// TensorPipe message.
-TORCH_API TensorpipeReadBuffers
-tensorpipeAllocate(tensorpipe::Message& tpMessage);
-
-// Convert a TensorPipe message back into an RPC message. This requires the data
-// to be available and can thus only be performed once the asynchronous read has
-// completed. The holder can be destroyed once this function returns.
-TORCH_API Message tensorpipeDeserialize(
-    tensorpipe::Message&& tpMessage,
-    TensorpipeReadBuffers&& holder);
 
 // Some Tensors are effectively views of larger Tensors, where only a small
 // subset of the Storage data is referenced. This normally is good and avoids
